Move hyperopt-loss functions to their own package
