+++ freqtrade/optimize/hyperopt_loss/hyperopt_loss_max_drawdown.py
