+++ freqtrade/optimize/hyperopt_loss/hyperopt_loss_sharpe.py
