+++ freqtrade/optimize/hyperopt_loss/hyperopt_loss_sortino_daily.py
